Modularize notifications management, add hooks, and improve UI
- Created `use-notifications` hook for managing notification state and badge updates.
- Extracted push subscriptions logic into `use-push-subscriptions` for reusability.
- Implemented `NotificationsBell` component for displaying unread badge.
- Added `NotificationsPanel` component for improved notification list presentation.
- Refactored service worker registration into `use-service-worker-registration` hook.
- Centralized types in `utils/types.tsx` for better type management.
- Enhanced UI with cleaner structure and component-based rendering.
