+++ utils/types.tsx
@@ -1,0 +1,3 @@
+export type SubsResponse = { count: number; subscriptions: unknown[] };
+
+export type NotifItem = { id: string; title: string; body: string; url?: string; timestamp: number };
